--- src/modules/auth/ResetPassword/schema/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./ResetPassword.schema";
-export * from "./ResetPassword.schema";
