--- src/modules/auth/ResetPassword/services/types/fetchPostResetPasswordService.types.ts
@@ -1,6 +1,0 @@
-import { IResetPasswordForm } from "modules/auth/ResetPassword/types";
-
-export interface IFetchPostResetPasswordService {
-	body: IResetPasswordForm;
-	token: string | undefined;
-}
