+++ src/modules/auth/ResetPassword/services/types/index.ts
@@ -1,1 +1,1 @@
-export * from "./fetchPostResetPasswordService.types";
+export * from "./post-reset-password.types";
