+++ src/modules/auth/ResetPassword/services/types/post-reset-password.types.ts
@@ -1,0 +1,6 @@
+import type { IResetPasswordForm } from "modules/auth/ResetPassword/types";
+
+export interface IPostResetPasswordService {
+	body: IResetPasswordForm;
+	token: string | undefined;
+}
