--- src/modules/auth/ResetPassword/styles/types/index.ts
@@ -1,3 +1,0 @@
-export interface ILinkStyle {
-    disabled: boolean;
-}
