+++ src/modules/auth/ResetPassword/types/index.ts
@@ -1,2 +1,1 @@
-export * as default from "./ResetPassword.types";
-export * from "./ResetPassword.types";
+export * from "./reset-password.types";
