+++ src/modules/auth/ResetPassword/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./define-subtitle-message";
+export * from "./define-title-message";
